requirements.txt
streamlit
scikit-learn
pandas
python-docx